Fix > code generation and add support for break statements
Fix > code generation and add support for break statements. Code
generation for the > operator was missing the jump after evaluation,
resulting it in always evaluating to negative.

Add support for break statements. Add an additional field to the
generators state record, tracking the end label of the current loop. Add
basic while loop with break test.